Moving temperature-aware-profiles algorithm stuff in to new file 'temperature_aware_profiles.h'.
